requirements.txt
# Timezone database — required on Windows for zoneinfo (ICT time-window filter)
tzdata>=2024.1

# Core bot dependencies
ccxt>=4.4.0
pandas>=2.2.0
numpy>=1.26.0
python-dotenv>=1.0.1
scikit-learn>=1.4.0

# ---------------------------------------------------------------------------
# To use your forked CCXT (https://github.com/htownz/ccxt):
#   pip uninstall ccxt -y
#   pip install git+https://github.com/htownz/ccxt.git@master#egg=ccxt
#
# Keep in sync with upstream by periodically running:
#   git -C /path/to/htownz-ccxt fetch upstream && git merge upstream/master
# ---------------------------------------------------------------------------

# Optional ML backends (install as needed)
# xgboost>=2.0.0
# lightgbm>=4.0.0

# Optional dashboard
# streamlit>=1.32.0
# plotly>=5.20.0

# Reinforcement learning (PPO agent)
# PyTorch is a transitive dependency of stable-baselines3 (~700 MB first install)
stable-baselines3>=2.3.0
gymnasium>=0.29.0

# Historical data backfill (no API key required)
yfinance>=0.2.0

# Hyperparameter tuning (python -m hogan_bot.rl_tune)
optuna>=3.0.0

# CoinGecko market intelligence (uses stdlib urllib — no extra package needed)
# Set COINGECKO_KEY in .env, then run:
#   python -m hogan_bot.fetch_coingecko --backfill   # one-time historical fill
#   python -m hogan_bot.fetch_coingecko               # daily cron

# GPR Index Excel readers (fetch_gpr.py uses xlrd for .xls, openpyxl for .xlsx)
xlrd>=2.0.1
openpyxl>=3.1.0

# Alternative / sentiment data (all use stdlib urllib — no extra packages needed)
# Free:
#   python -m hogan_bot.fetch_feargreed --backfill   # Alternative.me Fear & Greed
#   python -m hogan_bot.fetch_news_sentiment         # CryptoPanic (set CRYPTOPANIC_KEY)
#   python -m hogan_bot.fetch_gpr                   # Caldara & Iacoviello GPR Index
# Paid ($78-148/mo):
#   python -m hogan_bot.fetch_glassnode --days 365  # Glassnode (set GLASSNODE_KEY)
#   python -m hogan_bot.fetch_santiment --days 365  # Santiment (set SANTIMENT_KEY)

# Metrics / monitoring (optional)
prometheus-client>=0.20.0

# ---------------------------------------------------------------------------
# Phase 2 — Event-driven architecture (WebSocket)
# ---------------------------------------------------------------------------
# ccxt[pro] or ccxt.pro for WebSocket candle streams
# Install with: pip install "ccxt[pro]"

# ---------------------------------------------------------------------------
# Phase 3 — MLflow experiment tracking
# ---------------------------------------------------------------------------
mlflow>=2.11.0

# ---------------------------------------------------------------------------
# Phase 4 — Online / continuous learning
# ---------------------------------------------------------------------------
river>=0.21.0              # optional: river online learning library
scipy>=1.12.0              # KL divergence for feature drift monitor

# ---------------------------------------------------------------------------
# Phase 5 — MCP server
# ---------------------------------------------------------------------------
fastapi>=0.110.0
uvicorn>=0.29.0

# ---------------------------------------------------------------------------
# Phase 6 — OpenBB macro data
# ---------------------------------------------------------------------------
# openbb>=4.3.0           # optional, large install; yfinance fallback always works

# ---------------------------------------------------------------------------
# Phase 7 — Production security
# ---------------------------------------------------------------------------
cryptography>=42.0.0      # AES-256 API key encryption at rest

# ---------------------------------------------------------------------------
# Phase 8 — LLM Intelligence Layer
# ---------------------------------------------------------------------------
# chromadb>=0.5.0          # local vector store for RAG (optional, ~500 MB)
# sentence-transformers>=3.0.0  # embeddings for RAG (optional, requires torch)
# openai>=1.0.0            # optional: for GPT-4o-mini trade explanations
# httpx>=0.27.0            # optional: async HTTP for Ollama API
